refactored image functionality to support images in tables
Created ImageManager class to keep track of all images, and write media objects to the zip at the end to prevent corruption.

ImageManager is also responsible for writing new images to the zip

Also created a RelationshipsManager class for creating new relationships (e.g. for new images) as well as keeping track of existing relationships